Moved files into a 7-in-1 sass structure
