+++ src/main.css
@@ -1,0 +1,17 @@
+/* Error: Can't find stylesheet to import.
+ *   ,
+ * 6 | 'abstracts/variables';
+ *   | ^^^^^^^^^^^^^^^^^^^^^
+ *   '
+ *   main.scss 6:1  root stylesheet */
+
+body::before {
+  font-family: "Source Code Pro", "SF Mono", Monaco, Inconsolata, "Fira Mono",
+      "Droid Sans Mono", monospace, monospace;
+  white-space: pre;
+  display: block;
+  padding: 1em;
+  margin-bottom: 1em;
+  border-bottom: 2px solid black;
+  content: "Error: Can't find stylesheet to import.\a   \2577 \a 6 \2502  'abstracts/variables';\a   \2502  ^^^^^^^^^^^^^^^^^^^^^\a   \2575 \a   main.scss 6:1  root stylesheet";
+}
