--- src/styles/main.scss
@@ -1,2 +1,0 @@
-//This file will inherit all other non-compiled scss files into one executable file.
-@import "reset";
